// src/models/doaModel.ts
export interface DoadoaModel {
  arab: string;
  indo: string;
  judul: string;
  source: string;
}

export interface DoatahlilModel {
  id: number;
  title: string;
  arabic: string;
  translation: string;
}

export interface DoaharianModel {
  title: string;
  arabic: string;
  latin: string;
  translation: string;
}